backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
gunicorn==22.0.0
python-dotenv==1.0.0
pydantic==2.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
python-multipart==0.0.6
cors==1.0.1
reportlab==4.2.0
